Flexible invoice transmission
