+++ src/pretix/base/models/orders.py
@@ -3293,6 +3293,9 @@ class InvoiceAddress(models.Model):
         blank=True
     )
 
+    transmission_type = models.CharField(max_length=255, default="email")
+    transmission_info = models.JSONField(null=True, blank=True)
+
     objects = ScopedManager(organizer='order__event__organizer')
     profiles = ScopedManager(organizer='customer__organizer')
